requirements.txt
python-dotenv==1.0.0
qbittorrent_api==2023.6.50
Requests==2.31.0
